fix(windows): wire Authenticode through config, verify packaged bytes, pin NSIS shape
Release-engineering fixes surfaced by a full recon of the Windows chain
(none of it has ever executed end to end):

- release.yml: tauri-cli 2.x reads the Authenticode thumbprint ONLY from
  bundle.windows.certificateThumbprint; the TAURI_WINDOWS_CERTIFICATE_
  THUMBPRINT env var is a v1 relic it never reads. Injected via --config.
- release.yml: verify-package now checks the bytes INSIDE the MSI
  (-Archive) - the bundler restores target/release/cc-reminder.exe to an
  unsigned snapshot, so the old gate both false-failed and never validated
  the installer contents. setup.exe/msi sidecars validated via
  -PublishedFile.
- tauri.conf.json: pin bundle.windows.nsis (currentUser, en+SimpChinese,
  language selector) instead of silently relying on defaults; sha256+RFC3161
  timestamp defaults made explicit.
- ci.yml: windows-2022 leg now runs clippy -D warnings and cargo test --lib
  so cfg(windows) code paths are linted and executed per-PR.
- clippy: fix 3 warnings that only exist under the windows target.
- i18n: platform-neutral permission-trigger suggested prompt (ls -la / is
  meaningless under cmd/PowerShell).

+++ src-tauri/src/security/permissions.rs
@@ -420,7 +420,7 @@ mod windows_permissions {
     }
 
     fn well_known_sid(kind: i32) -> Result<Vec<u8>, AppError> {
-        let mut size = SECURITY_MAX_SID_SIZE as u32;
+        let mut size = SECURITY_MAX_SID_SIZE;
         let mut sid = vec![0_u8; size as usize];
         if unsafe { CreateWellKnownSid(kind, null_mut(), sid.as_mut_ptr().cast(), &mut size) } == 0
         {
